# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ChatBot Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/initiate-platform:
    post:
      operationId: initiatePlatformPayment
      tags: [payments]
      summary: Initiate Ksh 50 M-Pesa STK push to unlock platform access
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInitiateInput"
      responses:
        "200":
          description: STK push sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentInitiateResult"
  /payments/initiate-bot:
    post:
      operationId: initiateBotPayment
      tags: [payments]
      summary: Initiate Ksh 20 M-Pesa STK push to unlock a specific bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotPaymentInput"
      responses:
        "200":
          description: STK push sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentInitiateResult"
  /payments/callback:
    post:
      operationId: mpesaCallback
      tags: [payments]
      summary: M-Pesa payment callback (called by Safaricom)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MpesaCallbackPayload"
      responses:
        "200":
          description: Acknowledged
  /payments/status:
    get:
      operationId: fetchPaymentStatus
      tags: [payments]
      summary: Check payment status for a user
      parameters:
        - in: query
          name: phone
          required: true
          schema:
            type: string
        - in: query
          name: botId
          required: false
          schema:
            type: string
        - in: query
          name: checkoutRequestId
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Payment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentStatus"
  /payments/history:
    get:
      operationId: fetchPaymentHistory
      tags: [payments]
      summary: Get paginated transaction history for a user
      parameters:
        - in: query
          name: phone
          required: true
          schema:
            type: string
        - in: query
          name: page
          required: false
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Transaction history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentHistoryResult"
...
components:
  schemas:
...
    PaymentInitiateInput:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
    BotPaymentInput:
      type: object
      required: [phone, botId]
      properties:
        phone:
          type: string
        botId:
          type: string
    MpesaCallbackPayload:
      type: object
      properties:
        Body:
          type: object
    PaymentInitiateResult:
      type: object
      required: [message, checkoutRequestId]
      properties:
        message:
          type: string
        checkoutRequestId:
          type: string
        merchantRequestId:
          type: string
        transactionRef:
          type: string
    PaymentStatus:
      type: object
      required: [platformUnlocked, unlockedBots, paymentStatus]
      properties:
        platformUnlocked:
          type: boolean
        unlockedBots:
          type: array
          items:
            type: string
        paymentStatus:
          type: string
          enum: [pending, completed, failed, cancelled, timeout, none]
        resultCode:
          type: string
        resultDesc:
          type: string
        mpesaReceiptNumber:
          type: string
        transactionRef:
          type: string
        merchantRequestId:
          type: string
        amount:
          type: number
        completedAt:
          type: string
        sessionToken:
          type: string
    TransactionRecord:
      type: object
      required: [_id, type, amount, amountCents, status, checkoutRequestId, createdAt]
      properties:
        _id:
          type: string
        type:
          type: string
          enum: [platform, bot]
        botId:
          type: string
          nullable: true
        amount:
          type: number
        amountCents:
          type: integer
        status:
          type: string
          enum: [pending, completed, failed, cancelled, timeout]
        transactionRef:
          type: string
          nullable: true
        checkoutRequestId:
          type: string
        merchantRequestId:
          type: string
          nullable: true
        mpesaReceiptNumber:
          type: string
          nullable: true
        mpesaTransactionDate:
          type: string
          nullable: true
        resultCode:
          type: string
          nullable: true
        resultDesc:
          type: string
          nullable: true
        completedAt:
          type: string
          nullable: true
        createdAt:
          type: string
    PaymentHistoryResult:
      type: object
      required: [transactions, total, page, limit, totalPages]
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/TransactionRecord"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
...