# lib/api-zod/src/generated/types/transactionRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChatBot Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionRecordStatus } from "./transactionRecordStatus";
import type { TransactionRecordType } from "./transactionRecordType";

export interface TransactionRecord {
  _id: string;
  type: TransactionRecordType;
  botId?: string | null;
  amount: number;
  amountCents: number;
  status: TransactionRecordStatus;
  transactionRef?: string | null;
  checkoutRequestId: string;
  merchantRequestId?: string | null;
  mpesaReceiptNumber?: string | null;
  mpesaTransactionDate?: string | null;
  resultCode?: string | null;
  resultDesc?: string | null;
  completedAt?: string | null;
  createdAt: string;
}